Add files via upload
The data consists of 14 attributes. Some attributes are continuous, like age and cholesterol level. Some are categorical, like sex and type of chest pain. The task is to use the first 13 attributes to predict the 14th - the presence of heart disease in a patient. This task is complicated by the fact that the training set is so small. We have to be very careful because even a modest (by today's standards) neural network has more degrees of freedom than there are data points in our set!

In this project, we'll explore various ML models for this task. Our main tools will be the python scikit-learn/